+++ ui/pages/details/templates/bridge-details/hooks.test.ts
@@ -1,7 +1,6 @@
 import { renderHook } from '@testing-library/react-hooks';
 import type { BridgeHistoryItem } from '@metamask/bridge-status-controller';
 import { selectBridgeHistoryItemByHash } from '../../../../ducks/bridge-status/selectors';
-import { getBridgeHistoryTokens } from './utils';
 import { useBridgeHistoryItem, useHistoryTokens } from './hooks';
 
 jest.mock('react-redux', () => ({
@@ -13,16 +12,31 @@ jest.mock('../../../../ducks/bridge-status/selectors', () => ({
   selectBridgeHistoryItemByHash: jest.fn(),
 }));
 
-jest.mock('./utils', () => ({
-  getBridgeHistoryTokens: jest.fn(),
-}));
-
 const mockSelectBridgeHistoryItemByHash =
   selectBridgeHistoryItemByHash as unknown as jest.Mock;
-const mockGetBridgeHistoryTokens =
-  getBridgeHistoryTokens as unknown as jest.Mock;
 
-const historyItem = { account: '0xabc' } as unknown as BridgeHistoryItem;
+const buildHistoryItem = (
+  overrides: Partial<BridgeHistoryItem> = {},
+): BridgeHistoryItem =>
+  ({
+    account: '0xabc',
+    quote: {
+      srcTokenAmount: '1000',
+      srcAsset: {
+        assetId: 'eip155:1/slip44:60',
+        decimals: 18,
+        symbol: 'ETH',
+      },
+      destTokenAmount: '2000',
+      destAsset: {
+        assetId: 'eip155:137/erc20:0xdai',
+        decimals: 18,
+        symbol: 'DAI',
+      },
+    },
+    status: { destChain: { amount: '2100' } },
+    ...overrides,
+  }) as unknown as BridgeHistoryItem;
 
 describe('useBridgeHistoryItem', () => {
   beforeEach(() => {
@@ -30,6 +44,7 @@ describe('useBridgeHistoryItem', () => {
   });
 
   it('returns the bridge history item for the given source tx hash', () => {
+    const historyItem = buildHistoryItem();
     mockSelectBridgeHistoryItemByHash.mockReturnValue(historyItem);
 
     const { result } = renderHook(() => useBridgeHistoryItem('0xhash'));
@@ -54,23 +69,39 @@ describe('useHistoryTokens', () => {
     jest.clearAllMocks();
   });
 
-  it('maps the resolved history item to tokens', () => {
-    const tokens = {
-      sourceToken: { symbol: 'ETH' },
-      destinationToken: { symbol: 'DAI' },
-    };
-    mockSelectBridgeHistoryItemByHash.mockReturnValue(historyItem);
-    mockGetBridgeHistoryTokens.mockReturnValue(tokens);
+  it('maps the history item to source and destination tokens', () => {
+    mockSelectBridgeHistoryItemByHash.mockReturnValue(buildHistoryItem());
 
     const { result } = renderHook(() => useHistoryTokens('0xhash'));
 
-    expect(mockGetBridgeHistoryTokens).toHaveBeenCalledWith(historyItem);
-    expect(result.current).toBe(tokens);
+    expect(result.current?.sourceToken).toStrictEqual({
+      amount: '1000',
+      assetId: 'eip155:1/slip44:60',
+      decimals: 18,
+      direction: 'out',
+      symbol: 'ETH',
+    });
+    expect(result.current?.destinationToken).toStrictEqual({
+      amount: '2100',
+      assetId: 'eip155:137/erc20:0xdai',
+      decimals: 18,
+      direction: 'in',
+      symbol: 'DAI',
+    });
+  });
+
+  it('falls back to the quote destination amount when the destination chain amount is missing', () => {
+    mockSelectBridgeHistoryItemByHash.mockReturnValue(
+      buildHistoryItem({ status: {} } as Partial<BridgeHistoryItem>),
+    );
+
+    const { result } = renderHook(() => useHistoryTokens('0xhash'));
+
+    expect(result.current?.destinationToken.amount).toBe('2000');
   });
 
   it('returns undefined when there is no history item', () => {
     mockSelectBridgeHistoryItemByHash.mockReturnValue(undefined);
-    mockGetBridgeHistoryTokens.mockReturnValue(undefined);
 
     const { result } = renderHook(() => useHistoryTokens('0xhash'));
 
@@ -78,8 +109,7 @@ describe('useHistoryTokens', () => {
   });
 
   it('memoizes the result while the history item is unchanged', () => {
-    mockSelectBridgeHistoryItemByHash.mockReturnValue(historyItem);
-    mockGetBridgeHistoryTokens.mockReturnValue({ sourceToken: {} });
+    mockSelectBridgeHistoryItemByHash.mockReturnValue(buildHistoryItem());
 
     const { result, rerender } = renderHook(() => useHistoryTokens('0xhash'));
     const firstResult = result.current;
@@ -87,29 +117,32 @@ describe('useHistoryTokens', () => {
     rerender();
 
     expect(result.current).toBe(firstResult);
-    expect(mockGetBridgeHistoryTokens).toHaveBeenCalledTimes(1);
   });
 
   it('recomputes when the history item changes', () => {
-    const nextHistoryItem = {
-      account: '0xdef',
-    } as unknown as BridgeHistoryItem;
-    const firstTokens = { sourceToken: { symbol: 'ETH' } };
-    const secondTokens = { sourceToken: { symbol: 'MATIC' } };
-
     mockSelectBridgeHistoryItemByHash
-      .mockReturnValueOnce(historyItem)
-      .mockReturnValueOnce(nextHistoryItem);
-    mockGetBridgeHistoryTokens
-      .mockReturnValueOnce(firstTokens)
-      .mockReturnValueOnce(secondTokens);
+      .mockReturnValueOnce(buildHistoryItem())
+      .mockReturnValueOnce(
+        buildHistoryItem({
+          quote: {
+            srcTokenAmount: '5',
+            srcAsset: { assetId: 'eip155:1/slip44:60', decimals: 18, symbol: 'ETH' },
+            destTokenAmount: '9',
+            destAsset: {
+              assetId: 'eip155:10/erc20:0xusdc',
+              decimals: 6,
+              symbol: 'USDC',
+            },
+          },
+        } as unknown as Partial<BridgeHistoryItem>),
+      );
 
     const { result, rerender } = renderHook(() => useHistoryTokens('0xhash'));
-    expect(result.current).toBe(firstTokens);
+    const firstResult = result.current;
 
     rerender();
 
-    expect(mockGetBridgeHistoryTokens).toHaveBeenCalledTimes(2);
-    expect(result.current).toBe(secondTokens);
+    expect(result.current).not.toBe(firstResult);
+    expect(result.current?.destinationToken.symbol).toBe('USDC');
   });
 });
