+++ ui/pages/details/templates/bridge-details/hooks.ts
@@ -2,7 +2,7 @@ import { useMemo } from 'react';
 import { useSelector } from 'react-redux';
 import { selectBridgeHistoryItemByHash } from '../../../../ducks/bridge-status/selectors';
 import type { MetaMaskReduxState } from '../../../../store/store';
-import { getBridgeHistoryTokens } from './utils';
+import type { TokenAmount } from '../../../../../shared/lib/activity/types';
 
 export const useBridgeHistoryItem = (sourceTxHash?: string) => {
   return useSelector((state) =>
@@ -12,11 +12,33 @@ export const useBridgeHistoryItem = (sourceTxHash?: string) => {
   );
 };
 
-export const useHistoryTokens = (sourceTxHash?: string) => {
+export const useHistoryTokens = (
+  sourceTxHash?: string,
+): { sourceToken: TokenAmount; destinationToken: TokenAmount } | undefined => {
   const bridgeHistoryItem = useBridgeHistoryItem(sourceTxHash);
 
-  return useMemo(
-    () => getBridgeHistoryTokens(bridgeHistoryItem),
-    [bridgeHistoryItem],
-  );
+  return useMemo(() => {
+    if (!bridgeHistoryItem) {
+      return undefined;
+    }
+
+    const { quote, status } = bridgeHistoryItem;
+
+    return {
+      sourceToken: {
+        amount: quote.srcTokenAmount,
+        assetId: quote.srcAsset.assetId,
+        decimals: quote.srcAsset.decimals,
+        direction: 'out',
+        symbol: quote.srcAsset.symbol,
+      },
+      destinationToken: {
+        amount: status.destChain?.amount ?? quote.destTokenAmount,
+        assetId: quote.destAsset.assetId,
+        decimals: quote.destAsset.decimals,
+        direction: 'in',
+        symbol: quote.destAsset.symbol,
+      },
+    };
+  }, [bridgeHistoryItem]);
 };
